requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements.txt -r requirements-dev.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-xdist==3.6.1  # Run tests in parallel
factory-boy==3.3.1  # Test data factories
faker==33.1.0  # Generate fake data for tests

# Code Quality
ruff==0.8.5  # Fast linter and formatter (replaces flake8, black, isort, and more)

# Terminal UI
rich==13.9.4  # Beautiful terminal formatting and output

# Environment Management
python-dotenv==1.0.1  # Load environment variables from .env file

# Type Checking
mypy==1.14.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1
types-redis==4.6.0.20241004

# Pre-commit Hooks
pre-commit==4.0.1

# Security
bandit[toml]==1.8.0  # Security issue detection

# Development Tools
ipython==8.30.0  # Better Python shell
django-extensions==3.2.3  # Additional management commands
werkzeug==3.1.3  # Better debugger for runserver_plus
django-debug-toolbar==4.4.6  # Debug toolbar for development

# Documentation (optional)
# sphinx==8.1.3
# sphinx-rtd-theme==3.0.2
